refactor: move getCells function to utils
